Move atomics to concurrency
